Addressed comments and extracted read metadata out of KSIS
