+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/stats/HostLevelIngestionStats.java
@@ -132,16 +132,6 @@ public class HostLevelIngestionStats extends AbstractVeniceStats {
   private final LongAdderRateGauge totalTombstoneCreationDCRRate;
 
   /**
-   * Measure the number of time request based metadata endpoint was invoked
-   */
-  private final Sensor requestBasedMetadataInvokeCount;
-
-  /**
-   * Measure the number of time request based metadata endpoint failed to respond
-   */
-  private final Sensor requestBasedMetadataFailureCount;
-
-  /**
    * @param totalStats the total stats singleton instance, or null if we are constructing the total stats
    */
   public HostLevelIngestionStats(
@@ -416,18 +406,6 @@ public class HostLevelIngestionStats extends AbstractVeniceStats {
         totalStats,
         () -> totalStats.leaderIngestionActiveActiveDeleteLatencySensor,
         avgAndMax());
-
-    this.requestBasedMetadataInvokeCount = registerPerStoreAndTotalSensor(
-        "request_based_metadata_invoke_count",
-        totalStats,
-        () -> totalStats.requestBasedMetadataInvokeCount,
-        new Rate());
-
-    this.requestBasedMetadataFailureCount = registerPerStoreAndTotalSensor(
-        "request_based_metadata_failure_count",
-        totalStats,
-        () -> totalStats.requestBasedMetadataFailureCount,
-        new Rate());
   }
 
   /** Record a host-level byte consumption rate across all store versions */
@@ -590,13 +568,5 @@ public class HostLevelIngestionStats extends AbstractVeniceStats {
 
   public void recordOffsetRegressionDCRError() {
     totalOffsetRegressionDCRErrorRate.record();
-  }
-
-  public void recordRequestBasedMetadataInvokeCount() {
-    requestBasedMetadataInvokeCount.record();
-  }
-
-  public void recordRequestBasedMetadataFailureCount() {
-    requestBasedMetadataFailureCount.record();
   }
 }
