+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/storage/IngestionMetadataRetriever.java
@@ -1,21 +1,15 @@
 package com.linkedin.davinci.storage;
 
 import com.linkedin.davinci.listener.response.AdminResponse;
-import com.linkedin.davinci.listener.response.MetadataResponse;
-import com.linkedin.davinci.listener.response.ServerCurrentVersionResponse;
 import com.linkedin.davinci.listener.response.TopicPartitionIngestionContextResponse;
 import com.linkedin.venice.utils.ComplementSet;
 import java.nio.ByteBuffer;
 
 
-public interface MetadataRetriever {
+public interface IngestionMetadataRetriever {
   ByteBuffer getStoreVersionCompressionDictionary(String topicName);
 
   AdminResponse getConsumptionSnapshots(String topicName, ComplementSet<Integer> partitions);
-
-  MetadataResponse getMetadata(String storeName);
-
-  ServerCurrentVersionResponse getCurrentVersionResponse(String storeName);
 
   TopicPartitionIngestionContextResponse getTopicPartitionIngestionContext(
       String versionTopic,
